change user/create rest to post

+++ src/main/java/com/messenger/cryptosha/dto/KeyDTO.java
@@ -3,17 +3,17 @@ package com.messenger.cryptosha.dto;
 import com.andreyka.crypto.ECPoint;
 
 public class KeyDTO {
-    Integer userId;
+    Long userId;
     ECPoint publicKey;
 
     public KeyDTO() {
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
